pyproject.toml
[project]
name = "phish-analyzer"
version = "0.4.3"
description = "Email phishing analysis tool"
authors = [
    { name="DC", email="dacraven95@gmail.com" }
]
requires-python = ">=3.10"

dependencies = [
    "dnspython>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.7",
    "reportlab>=4.0.0",
    "halo>=0.0.31",
    "requests>=2.32.0",
    "playwright>=1.43,<2.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["phish_analyzer"]

[project.optional-dependencies]
dev = [
    "pylint>=3.0.0",
]
